requirements.txt
streamlit
langchain
langchain-community
openai
pypdf
python-dotenv
